chore: abstract MockBookmarksProvider
